skills/seed-observatories/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "observatory-seeder"
version = "1.0.0"
description = "Observatory seeding pipeline for Starview - discovers observatories from Wikidata with AI-validated images"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Starview", email = "hello@starview.app"}
]
requires-python = ">=3.11"
dependencies = [
    "requests>=2.28.0",
    "Pillow>=9.0.0",
]
keywords = ["observatory", "wikidata", "astronomy", "stargazing"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
]

[project.scripts]
observatory-seeder = "observatory_seeder.run:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["observatory_seeder*"]
